examples/mysql-pymysql/requirements.txt
dbop-core
pymysql>=1.1.0
python-dotenv
cryptography>=42
